Move Footer and Navigation components from components/Home directory to components directory.
